# docs/openapi/safe-zone.yaml
openapi: 3.1.0
info:
  title: Safe Zone API
  version: 0.1.0
  description: Local-first anti-phishing API for domain analysis, policy checks, telemetry, and administration.
servers:
  - url: https://{host}
    variables:
      host:
        default: safezone.example.com
  - url: http://localhost:8080
paths:
...
  /dns-query:
    get:
      summary: DNS-over-HTTPS query using RFC 8484 GET
      parameters:
        - name: dns
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DNS wire-format response
          content:
            application/dns-message:
              schema:
                type: string
                format: binary
    post:
      summary: DNS-over-HTTPS query using RFC 8484 POST
      requestBody:
        required: true
        content:
          application/dns-message:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: DNS wire-format response
          content:
            application/dns-message:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_session
    apiKeyAuth:
      type: http
      scheme: bearer
      description: Use `Bearer <SAFE_ZONE_ADMIN_API_KEY>`.
...